Add files via upload
Source code management
All source code is stored in TFS but only items required to build the solution should be checked in. 
Add third party components
When adding third party components you should use npm. It is recommended to add third party components to the solution in a separate branch and ones the components has been verified, merge back to the main branch.

When installing components use the following syntax:
npm install <component> --save
Add third party components
Source code management
Only the registration of the third party component should be stored in TFS, e.g. only the registration of the component in package.json file.
Exception
In some specific cases you need to store the third party files attached to the solution in order to guarantee future quality and maintenance. In these situations the files needs to be stored in the lib-folder (\Volvo.Perf.ClientUI\PerfCli\lib).
Upgrade third party components
In general the different third party components and their versions needs to be stable but on regular basis, the different third party components needs to be upgraded. An upgrade needs to be a planned activity and should be done early in the phase of developing a new version so that the combination of components provide required solution quality.
Branch strategy
In general all development is made referenced the Main branch in TFS. If a major change is made in the application such as architectural changes or a major refactoring, you need to branch from Main into a Dev-branch. Make the change and once it has been verified and accepted, merge back to Main-branch.
Build and deploy process
System documentation
All system documentation will be stored in the share point site allocated to the project: https://volvogroup.sharepoint.com/sites/proj-wis-perf-renewal-project/.
Requirements
The different documents that are used to describe the system are stored in the folder Documents/Requirements/PERF.
 

Instructions
In order to make it easier for new resources to become productive some how-to instructions are written. All these instructions are stored in the Wiki/How to folder.